Move axios GET requst into ShoppingListItem.js
